+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 20.11.5
+* Added new configuration fields to manipulate internal SDK value and key limits.
+* Warning! This release will introduce configurable maximum size limits for values and keys throughout the SDK. If they would exceed the limits, they would be truncated.
+* Fixed a bug that caused requests to arrive out of their intended order
+
 ## 20.11.4
 * Added "lock" checks for all public SDK methods.
 * Fixed race issue bug that occured when events are combined into a request.